[ID-33] pages reorganization
