+++ src/components/User/Reservations/reservations.js
@@ -7,7 +7,8 @@ const Reservations = () => {
   const location = useLocation();
 
   useEffect(() => {
-    if (location.state && location.state.data) console.log(location.state.data);
+    if (location.state && location.state.data)
+      console.log(location.state.data);
   }, []);
 
   return <div className="reservations-container">Reservations</div>;
