+++ src/screens/private/User/Reservation/reservation.js
@@ -1,8 +1,8 @@
 import { useEffect } from "react";
 import { auth } from "../../../../services/utils";
-
-import ReservationsComponent from "../../../../components/Reservations/reservations";
 import { useNavigate } from "react-router-dom";
+
+import ReservationsComponent from "../../../../components/User/Reservations/reservations";
 
 const Reservation = () => {
   const navigate = useNavigate();
